% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_filters.R
\name{adjacent_segments_filter}
\alias{adjacent_segments_filter}
\title{Runs the adjacent segments filter on each sample in the samples list.}
\usage{
adjacent_segments_filter(
  samples_list,
  comp_table,
  breakpoint_lookup_table,
  threshold = 0.6
)
}
\arguments{
\item{samples_list}{List of data.tables for each sample. Data tables should
already be marked with segment IDs of the initial segmentation.}

\item{comp_table}{Table comparing the segment means for each sample in the set.
Should contain entries for each sample in the samples list. Can be created using
`make_comparison_table(samples_list)`, and can be further filtered with
`pairwise_filter(comp_table)`.}

\item{breakpoint_lookup_table}{Table that allows for quick check of breakpoint
significance level. Create using `make_breakpoint_test_lookup_table(samples_list)`}

\item{threshold}{for effect size in breakpoint assessment}
}
\description{
Runs the adjacent segments filter on each sample in the samples list.
}
