% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{annotate_ascat_segments}
\alias{annotate_ascat_segments}
\title{Annotate `breaks` (=ascat_result$segments) so each segment is given two IDs:
  dataset_segment_id - uniquely identifies the segment (chr, startpos, endpos) within the entire dataset
                     - (a segment can occur in multiple samples)
  sample_segment_id - gives the segment an index within the sample
And two frequencies:
  breakpoint_frequency - number of times a given startpos is observed
  segment_frequency - number of times a unique segment is observed}
\usage{
annotate_ascat_segments(breaks)
}
\description{
Annotate `breaks` (=ascat_result$segments) so each segment is given two IDs:
  dataset_segment_id - uniquely identifies the segment (chr, startpos, endpos) within the entire dataset
                     - (a segment can occur in multiple samples)
  sample_segment_id - gives the segment an index within the sample
And two frequencies:
  breakpoint_frequency - number of times a given startpos is observed
  segment_frequency - number of times a unique segment is observed
}
