% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_distance_matrix_functions.R
\name{ascat_single_sample}
\alias{ascat_single_sample}
\title{Extract a single sample from an ascat data object}
\usage{
ascat_single_sample(ascat_data, samplename)
}
\description{
Extract a single sample from an ascat data object
}
