% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{ascatobj_to_datatable}
\alias{ascatobj_to_datatable}
\title{Convert an ascat object (not an ascat result) to a data table
containing SNP positions, logR and BAF at each position,
and segmented average LogR and BAF where available}
\usage{
ascatobj_to_datatable(ascat_obj, ascat_result = NULL)
}
\arguments{
\item{ascat_obj}{ASCAT data object, including segmentation information}

\item{ascat_result}{ASCAT analysis result object}
}
\value{
data table representation of ascat data
}
\description{
Convert an ascat object (not an ascat result) to a data table
containing SNP positions, logR and BAF at each position,
and segmented average LogR and BAF where available
}
