% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{beta_ab}
\alias{beta_ab}
\title{Find a and b s.t. Mean(Beta(a,b)) = mn and Var(Beta(a,b)) = va}
\usage{
beta_ab(mn, va)
}
\description{
Find a and b s.t. Mean(Beta(a,b)) = mn and Var(Beta(a,b)) = va
}
