% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution_helpers.R
\name{beta_boundary}
\alias{beta_boundary}
\title{Find boundary between two Beta distributions}
\usage{
beta_boundary(a1, b1, a2, b2, p1, p2, log = TRUE, plot = FALSE)
}
\description{
Find boundary between two Beta distributions
}
