% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_model.R
\name{bmm_likelihood}
\alias{bmm_likelihood}
\title{Overall log bmm_likelihood of the data given the mixture model}
\usage{
bmm_likelihood(m)
}
\arguments{
\item{m}{Matrix of component log bmm_likelihoods}
}
\description{
Overall log bmm_likelihood of the data given the mixture model
}
