% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_model.R
\name{bmm_ll_to_prob}
\alias{bmm_ll_to_prob}
\title{Convert log-bmm_likelihood matrix to probabilities}
\usage{
bmm_ll_to_prob(llmat)
}
\arguments{
\item{llmat}{- Matrix with log-bmm_likelihoods for each mixture component in columns}
}
\description{
Convert log-bmm_likelihood matrix to probabilities
}
