% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_model.R
\name{bmm_loglik}
\alias{bmm_loglik}
\title{Log bmm_likelihood of data under binomial distribution}
\usage{
bmm_loglik(param, data, binom_log_constant)
}
\arguments{
\item{param}{- Parameter of binomial distribution}

\item{data}{- Matrix of successes, failures in columns}

\item{binom_log_constant}{- Logarithm of the binomial constant for data}
}
\description{
Log bmm_likelihood of data under binomial distribution
}
