% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{calc_combined_genotype}
\alias{calc_combined_genotype}
\title{Calculates host genotype
host = hom if (coverage ≥ 10, as above) and 0 reads ref | alt
host = het if ≥ 3 reads for both ALT and REF & .25 ≤ BAF ≤ .75}
\usage{
calc_combined_genotype(h_ref, h_alt, t_ref, t_alt, baf_interval = c(0.25,
  0.75), min_reads = 3)
}
\arguments{
\item{data}{= data frame / table with columns 'h_ref', 'h_alt', 't_ref'
and 't_alt'}
}
\description{
Calculates host genotype
host = hom if (coverage ≥ 10, as above) and 0 reads ref | alt
host = het if ≥ 3 reads for both ALT and REF & .25 ≤ BAF ≤ .75
}
