% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clamp}
\alias{clamp}
\title{Clamp values in \code{v} to \code{lower} and \code{upper}, so that
any value \code{val} in \code{v} for which \code{val}
< \code{lower} is replaced by \code{lower}, and
\code{val} > \code{upper} is replaced by \code{upper}}
\usage{
clamp(v, lower, upper)
}
\arguments{
\item{v}{A vector of numbers.}

\item{lower}{A number. Any values in `v` below this are replaced.}

\item{upper}{A number. Any values in `v` above this are replaced.}
}
\value{
A vector like \code{v}, with all values less than
\code{lower} or greater than \code{upper} replaced.
}
\description{
Clamp values in \code{v} to \code{lower} and \code{upper}, so that
any value \code{val} in \code{v} for which \code{val}
< \code{lower} is replaced by \code{lower}, and
\code{val} > \code{upper} is replaced by \code{upper}
}
\examples{
clamp(c(0, 0.4, 0.6, 1.0), 0.001, 0.999)
# [1] 0.001 0.400 0.600 0.999
}
