% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_correction.R
\name{clip}
\alias{clip}
\title{Clip numeric vector so that its lowest element is >= low,
and its highest is <= high}
\usage{
clip(v, low, high)
}
\description{
Clip numeric vector so that its lowest element is >= low,
and its highest is <= high
}
