% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{convert_genotyped_breaks_to_intervals}
\alias{convert_genotyped_breaks_to_intervals}
\title{Helper function to convert a tall data.table of genotyped breakpoints
into segment intervals. ('Tall' means the data for each sample
is stacked vertically (melted), as opposed to horizontally, in
separate columns (dcasted).}
\usage{
convert_genotyped_breaks_to_intervals(
  breaks,
  snpdata,
  allow_pass_in_any_sample = FALSE
)
}
\arguments{
\item{breaks}{- data.table with columns 'chr', 'pos', 'sample' and 'PASS',
where (chr,pos) are the coordinates of a breakpoint, sample is the name
of the sample the breakpoint appears in, and 'PASS' is a boolean indicating
whether the breakpoint exists in the sample}

\item{snpdata}{- data.table with columns 'chr', 'pos', where (chr,pos) are
SNP coordinates}
}
\value{
- data.table of (chr,startpos,endpos) intervals, where startpos
and endpos are tied to available SNP positions from snpdata
}
\description{
Helper function to convert a tall data.table of genotyped breakpoints
into segment intervals. ('Tall' means the data for each sample
is stacked vertically (melted), as opposed to horizontally, in
separate columns (dcasted).
}
