% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{convert_to_het_host}
\alias{convert_to_het_host}
\title{Uses \code{deconvolute} to calculate adjusted tumour vaf as it would be if all host contamination were heterozygous}
\usage{
convert_to_het_host(tumour_ref, tumour_alt, host_ref, host_alt, logr, purity)
}
\description{
Uses \code{deconvolute} to calculate adjusted tumour vaf as it would be if all host contamination were heterozygous
}
