% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{count_matrix}
\alias{count_matrix}
\title{Construct a matrix of read counts given the marginal REF and ALT totals,
and the conditional probabilities that a read came from the host, given its
status as an A (ref) or a B (alt) read
Table:}
\usage{
count_matrix(ref, alt, p)
}
\arguments{
\item{ref}{= Integer count of REF reads observed in (contaminated) tumour sample}

\item{alt}{= Integer count of ALT reads observed in tumour sample}

\item{p}{= Vector of conditional probabilities [p(H|A), p(H|B)]}
}
\description{
|   T   |   H    | total
 ---|-------|--------|------
  A |   .   | p(H|A) |  REF
  B |   .   | p(H|B) |  ALT
 ---|-------|--------|------
    |Σ(.|T) | Σ(.|H) |REF+ALT
}
