% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Probability mass function for the beta-binomial distribution}
\usage{
dbetabinom(x, n, a, b)
}
\arguments{
\item{x}{Integer: Number of successes}

\item{n}{Integer: Number of trials}

\item{a}{Numeric: alpha parameter of the beta distribution}

\item{b}{Numeric: beta parameter of the beta distribution}
}
\value{
Numeric: probability of x given n, a, b
}
\description{
Probability mass function for the beta-binomial distribution
}
