% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_symmetrical_beta_mixture_distribution.R
\name{ddoublesbeta}
\alias{ddoublesbeta}
\title{Density function for symmetrical beta mixture}
\usage{
ddoublesbeta(x, a1, b1, a2, b2, w1, w2, w3, log = FALSE)
}
\description{
Density function for symmetrical beta mixture
}
