% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolute}
\alias{deconvolute}
\title{Estimate the host and tumour components from a host-contaminated tumour sample}
\usage{
deconvolute(tumour_ref, tumour_alt, host_ref, host_alt, logr, purity,
  plot = FALSE)
}
\arguments{
\item{tumour_ref}{number of REF reads observed in host-contaminated tumour, at current site}

\item{tumour_alt}{number of ALT reads observed in host-contaminated tumour, at current site}

\item{host_ref}{number of REF reads observed in matched host, at current site}

\item{host_alt}{number of ALT reads observed in matched host, at current site}

\item{logr}{log of ratio between tumour coverage and matched host coverage, at current site}

\item{purity}{Estimate of tumour purity (aka aberrant cell count)}

\item{plot}{Should a plot be drawn?}
}
\value{
A list with the following elements:
\itemize{
\item par = optimised parameters (\eqn{p(H | A)}, \eqn{p(H | B)})
\item cov = estimated covariance of optimised parameters
(Fisher information - inverse of Hessian)
\item counts = Matrix of inferred read counts, divided by Host/Tumour, A/B-allele
\item pure_tumour_vaf = Estimate of tumour VAF in absence of host
\item pure_host_vaf = Estimate of host VAF in absence of tumour
\item sample_vaf_with_het_host = An estimate of what the sample VAF would be if the host were exactly
heterozygous
}
}
\description{
Estimate the host and tumour components from a host-contaminated tumour sample
}
\examples{
deconvolute(90, 15, 71, 8, 0.03670243, 0.59, FALSE)
}
