% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{disjoin_ascat_segments}
\alias{disjoin_ascat_segments}
\title{Create a new table of segments that includes every breakpoint from every sample in
`datatable`, where datatable = ascat_result$segments}
\usage{
disjoin_ascat_segments(datatable)
}
\description{
Create a new table of segments that includes every breakpoint from every sample in
`datatable`, where datatable = ascat_result$segments
}
