% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrical_beta_mixture_distribution.R
\name{dsbeta}
\alias{dsbeta}
\title{Density function for symmetrical beta mixture}
\usage{
dsbeta(x, a, b, w, log = FALSE)
}
\description{
Density function for symmetrical beta mixture
}
