% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_symmetrical_beta_mixture_distribution.R
\name{est_doublesbeta}
\alias{est_doublesbeta}
\title{Estimate parameters of symmetrical Beta distribution by maximum likelihood
(assumes mixture weight fixed to 0.5)}
\usage{
est_doublesbeta(data, startval = c(1, 10, 1, 10, 0.5, 0.5, 0.5))
}
\description{
Estimate parameters of symmetrical Beta distribution by maximum likelihood
(assumes mixture weight fixed to 0.5)
}
