% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflected_beta_distribution.R
\name{est_reflbeta}
\alias{est_reflbeta}
\title{Estimate parameters of reflected Beta distribution by maximum likelihood}
\usage{
est_reflbeta(data, startval = c(1, 10))
}
\description{
Estimate parameters of reflected Beta distribution by maximum likelihood
}
