% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximisation_beta_mixture_model.R
\name{estimate_mixture}
\alias{estimate_mixture}
\title{Run Expectation Maximisation to optimise the fit of a mixture of Beta components
to data}
\usage{
estimate_mixture(
  data,
  init_params,
  init_pi,
  max_steps = 1000,
  tol = 1e-05,
  plot = FALSE
)
}
\arguments{
\item{data}{(numeric matrix) - Observations in the range [0,1]}

\item{init_params}{(numeric matrix) - Matrix of initial Beta(a, b) parameters, with
one row per mixture component.}

\item{init_pi}{(numeric vector) - Vector of initial mixture component proportions,
with one entry per mixture component.}

\item{max_steps}{(integer) - Maximum number of iterations to perform while optimising}

\item{tol}{(numeric) - Terminate optimisation when change in parameters between iterations
is less than tol}

\item{plot}{(bool) - Plot optimisation progress to active graphical device}
}
\value{
list(params, pi, usedsteps, assignment, uncertainty) - Optimised parameters and mixture proportions,
number of steps used, the assignment of each point to a mixture component,
and the degree of uncertainty for each point
}
\description{
Run Expectation Maximisation to optimise the fit of a mixture of Beta components
to data
}
