% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{estimate_purity}
\alias{estimate_purity}
\title{Rough estimate of purity}
\usage{
estimate_purity(host_vaf, tumour_vaf, logr, hhetl = 0.3, hhetu = 0.7,
  thoml = 0.25, thomu = 0.75, lrl = -0.75, lru = 0.75)
}
\description{
Rough estimate of purity
}
