% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{estimate_tumour_vaf}
\alias{estimate_tumour_vaf}
\title{Uses \code{deconvolute} to calculate adjusted tumour vaf as it would be if there were no host contamination}
\usage{
estimate_tumour_vaf(tumour_ref, tumour_alt, host_ref, host_alt, logr, purity)
}
\description{
Uses \code{deconvolute} to calculate adjusted tumour vaf as it would be if there were no host contamination
}
