% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{expectation_tumour_vaf}
\alias{expectation_tumour_vaf}
\title{Calculates the expectation of the tumour VAF, marginalized over
all valid contingency tables, conditioned on observed read totals.
Much slower than `fast_estimate_tumour_vaf`}
\usage{
expectation_tumour_vaf(
  total_readdepth,
  alt_readdepth,
  logr,
  host_total_readdepth,
  host_alt_readdepth,
  purity,
  ploidy,
  host_ploidy = 2,
  use_pseudocount = TRUE
)
}
\arguments{
\item{total_readdepth}{Integer: Observed total read depth in the tumour sample}

\item{alt_readdepth}{Integer: Observed number of reads carrying the alt allele in the tumour sample}

\item{logr}{Numeric: Estimated log2 of the ratio of tumour read depth/host read depth}

\item{host_total_readdepth}{Integer: Observed total read depth in the matched host sample}

\item{host_alt_readdepth}{Integer: Observed number of reads carrying the alt allele in the matched host sample}

\item{purity}{Numeric [0, 1]: Estimated aberrant cell fraction of the tumour sample}

\item{ploidy}{Numeric: Estimated ploidy of the average tumour cell}

\item{host_ploidy}{Numeric: Estimated ploidy of a normal cell (default=2)}

\item{use_pseudocount}{Logical: If TRUE, adds 0.5 to the alpha and beta parameters of the beta-binomial distribution}
}
\description{
Calculates the expectation of the tumour VAF, marginalized over
all valid contingency tables, conditioned on observed read totals.
Much slower than `fast_estimate_tumour_vaf`
}
