% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expected_loh_vaf}
\alias{expected_loh_vaf}
\title{Calculate the expected VAF for a SNP at the given purity, ploidy and logR,
if that SNP were LOH or otherwise homozygous in the founder
Can be used produce a segment threshold, above which a SNP is likely to have been
HET in the founder.}
\usage{
expected_loh_vaf(purity, logr, ploidy)
}
\arguments{
\item{purity}{Estimated tumour purity (real value [0..1])}

\item{logr}{Log (base-2) of the read depth ratio between tumour and normal (real value)}

\item{ploidy}{Estimated overall tumour ploidy (real value)}
}
\value{
The estimated VAF, had the original state of the SNP been homozygous
(real value [0..1]). NB: This is hypothetical, not a prediction.
}
\description{
Calculate the expected VAF for a SNP at the given purity, ploidy and logR,
if that SNP were LOH or otherwise homozygous in the founder
Can be used produce a segment threshold, above which a SNP is likely to have been
HET in the founder.
}
\examples{
expected_loh_vaf(0.817, c(1.713900, 1.607939, 1.629899, 1.570724, 1.666647), 2.114)
# [1] 0.02595942 0.02788265 0.02747302 0.02859042 0.02680059
}
