% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{extract_sample_name}
\alias{extract_sample_name}
\title{Helper function to get the sample name from a data.table
of tumour data}
\usage{
extract_sample_name(dt)
}
\description{
Helper function to get the sample name from a data.table
of tumour data
}
