% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{fast_estimate_contingency_table}
\alias{fast_estimate_contingency_table}
\title{Fast approximate estimate of pure_tumour_vaf from a mixed sample}
\usage{
fast_estimate_contingency_table(
  total_readdepth,
  alt_readdepth,
  logr,
  host_total_readdepth,
  host_alt_readdepth,
  purity,
  ploidy,
  host_ploidy = 2
)
}
\arguments{
\item{total_readdepth}{Tumour sample total read depth}

\item{alt_readdepth}{Tumour sample alt read depth}

\item{logr}{Tumour sample logR}

\item{host_total_readdepth}{Host sample total read depth}

\item{host_alt_readdepth}{Host sample alt read depth}

\item{purity}{Estimated purity of tumour sample}

\item{ploidy}{Estimated ploidy of the tumour sample}

\item{host_ploidy}{Estimated ploidy of the host sample (default = 2)}
}
\description{
Fast approximate estimate of pure_tumour_vaf from a mixed sample
}
