% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{fast_estimate_tumour_logodds}
\alias{fast_estimate_tumour_logodds}
\title{Fast approximate estimate of pure_tumour_vaf from a mixed sample}
\usage{
fast_estimate_tumour_logodds(
  total_readdepth,
  alt_readdepth,
  logr,
  hvaf,
  purity,
  pseudocount = 0.5
)
}
\arguments{
\item{total_readdepth}{Tumour sample total read depth}

\item{alt_readdepth}{Tumour sample alt read depth}

\item{logr}{Tumour sample logR}

\item{hvaf}{Host sample VAF}

\item{purity}{Estimated purity of tumour sample}

\item{pseudocount}{Avoid infinite log-odds by adding a pseudocount to each
cell of the contingency table. Default = 0.5 (same as Facets).}
}
\description{
Fast approximate estimate of pure_tumour_vaf from a mixed sample
}
