% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{fast_estimate_tumour_logodds_and_variance}
\alias{fast_estimate_tumour_logodds_and_variance}
\title{Fast approximate estimate of log-odds ratio, and its variance, from a mixed sample,
with the host converted to heterozygous diploid.}
\usage{
fast_estimate_tumour_logodds_and_variance(
  total_readdepth,
  alt_readdepth,
  logr,
  host_total_readdepth,
  host_alt_readdepth,
  purity,
  ploidy,
  host_ploidy = 2,
  pseudocount = 0.5
)
}
\arguments{
\item{total_readdepth}{Tumour sample total read depth}

\item{alt_readdepth}{Tumour sample alt read depth}

\item{logr}{Tumour sample logR}

\item{host_total_readdepth}{Host sample total read depth}

\item{host_alt_readdepth}{Host sample alt read depth}

\item{purity}{Estimated purity of tumour sample}

\item{ploidy}{Estimated ploidy of the tumour sample}

\item{host_ploidy}{Estimated ploidy of the host sample (default = 2)}

\item{pseudocount}{Avoid infinite log-odds by adding a pseudocount to each
cell of the contingency table. Default = 0.5 (same as Facets).}
}
\description{
Fast approximate estimate of log-odds ratio, and its variance, from a mixed sample,
with the host converted to heterozygous diploid.
}
