% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{fast_estimate_tumour_vaf}
\alias{fast_estimate_tumour_vaf}
\title{Fast approximate estimate of pure_tumour_vaf from a mixed sample}
\usage{
fast_estimate_tumour_vaf(total_readdepth, alt_readdepth, logr, hvaf, purity)
}
\arguments{
\item{total_readdepth}{Tumour sample total read depth}

\item{alt_readdepth}{Tumour sample alt read depth}

\item{logr}{Tumour sample logR}

\item{hvaf}{Host sample VAF}

\item{purity}{Estimated purity of tumour sample}
}
\description{
Fast approximate estimate of pure_tumour_vaf from a mixed sample
}
