% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{filter_segments_by_snplist}
\alias{filter_segments_by_snplist}
\title{Post-process a disjointed segments list to remove segments that cover 0 SNPs
Adds a column, Nsnps, the number of SNPs covered by each segment}
\usage{
filter_segments_by_snplist(snp_list, segments)
}
\arguments{
\item{snp_list}{(data.table) A list of SNPs with columns chr (chromosome),
startpos (SNP position) and endpos (=startpos). Other columns can be present,
but ignored.}

\item{segments}{(data.table) A list of segments with columns chr (chromosome),
startpos and endpos (start and end positions of the segments)}
}
\description{
Post-process a disjointed segments list to remove segments that cover 0 SNPs
Adds a column, Nsnps, the number of SNPs covered by each segment
}
