% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_distance_matrix_functions.R
\name{find_box_index}
\alias{find_box_index}
\title{Find peaks in the ASCAT sunrise plot data.}
\usage{
find_box_index(dm, psi, rho)
}
\arguments{
\item{dm}{= Distance matrix from ascat_result$distance_matrix}

\item{nbr_size}{= Neighbourhood size for peakfinding (e.g. 5x5 matrix centred on peak)}
}
\description{
Find peaks in the ASCAT sunrise plot data.
}
