% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_copynumber_to_data.R
\name{fit_minor_copy_number}
\alias{fit_minor_copy_number}
\title{Finds the best fitting minor copy number for the VAF values in `vaf_data_`,
given the `purity_` and `total_cn` values.}
\usage{
fit_minor_copy_number(vaf_data_, purity_, total_cn)
}
\description{
Finds the best fitting minor copy number for the VAF values in `vaf_data_`,
given the `purity_` and `total_cn` values.
}
