% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_copynumber_to_data.R
\name{fit_total_copy_number}
\alias{fit_total_copy_number}
\title{Finds the best fitting total copy number for the logR values in `logr_data_`,
given the `purity_` and `ploidy_` values.}
\usage{
fit_total_copy_number(logr_data_, purity_, ploidy_, max_cn_ = 10)
}
\description{
Finds the best fitting total copy number for the logR values in `logr_data_`,
given the `purity_` and `ploidy_` values.
}
