% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_correction.R
\name{gatk_correct}
\alias{gatk_correct}
\title{Do tangent normalisation on a tumour sample}
\usage{
gatk_correct(data, decomp, n_singular_values, verbose = TRUE)
}
\description{
Do tangent normalisation on a tumour sample
}
