% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_correction.R
\name{gatk_decompose}
\alias{gatk_decompose}
\title{Obtain read count normalisation info from a panel of normals}
\usage{
gatk_decompose(data, filter1 = 0.1, filter2 = 0.05, filter3 = 0.001,
  nsvs = 20, verbose = TRUE)
}
\arguments{
\item{data}{(data.table / data.frame) Table of read counts, with columns representing samples
and rows representing genomic locations or intervals}

\item{filter1}{(real) Filter out rows where the median value is below the quantile threshold
given by this parameter (default = 0.1)}

\item{filter2}{(real) Filter out rows where more than this threshold proportion of samples have
zero read coverage (default = 0.05)}

\item{filter3}{(real) Clamp (Winsorize) values to the quantiles given by this value (and 1 - this value)}

\item{nsvs}{(int) Number of Singular Values to retain (cannot exceed number of samples)}
}
\description{
Obtain read count normalisation info from a panel of normals
}
