% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_correction.R
\name{gatk_gc_correct}
\alias{gatk_gc_correct}
\title{GATK-style GC-correction of read count information}
\usage{
gatk_gc_correct(
  read_counts,
  gc_content,
  num_bins = 101,
  correlation_length = 0.02
)
}
\description{
GATK-style GC-correction of read count information
}
