% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximisation_beta_mixture_model.R
\name{get_delta}
\alias{get_delta}
\title{Return the largest scaled absolute difference among all parameter
values in a comparison of old and new values}
\usage{
get_delta(new_params, old_params, new_pi, old_pi)
}
\description{
Return the largest scaled absolute difference among all parameter
values in a comparison of old and new values
}
