% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximisation_beta_mixture_model.R
\name{get_init}
\alias{get_init}
\title{Get initial parameter estimates for the mixture model}
\usage{
get_init(v, k = 3)
}
\arguments{
\item{v}{Model data}

\item{k}{Number of clusters}
}
\value{
(List) Initial estimates of params and pi
}
\description{
Get initial parameter estimates for the mixture model
}
