% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logr_computation.R
\name{get_median_host}
\alias{get_median_host}
\title{Revised logr Calculation
2 steps: 1 - get host median coverage data
         2 - compute logr based on median host coverage}
\usage{
get_median_host(m)
}
\description{
Revised logr Calculation
2 steps: 1 - get host median coverage data
         2 - compute logr based on median host coverage
}
