% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{get_updated_params}
\alias{get_updated_params}
\title{EM - MAXIMISATION STEP
Given data and a matrix of mixture component weights for each
data point, estimate new parameter values (params) and mixture
proportions (pi)}
\usage{
get_updated_params(data, weights)
}
\description{
EM - MAXIMISATION STEP
Given data and a matrix of mixture component weights for each
data point, estimate new parameter values (params) and mixture
proportions (pi)
}
