% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{get_weights}
\alias{get_weights}
\title{EM - EXPECTATION STEP
Given data, a matrix of parameters (params) and a vector of
mixture component proportions (pi), compute the probability
that each data point came from each component (weights)}
\usage{
get_weights(data, params, pi)
}
\description{
EM - EXPECTATION STEP
Given data, a matrix of parameters (params) and a vector of
mixture component proportions (pi), compute the probability
that each data point came from each component (weights)
}
