% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{hmm_segmentation}
\alias{hmm_segmentation}
\title{Segment copy number using an HMM.}
\usage{
hmm_segmentation(cn, prob01, prob10, mean0, mean1, minrun = 1000, python_exe,
  return_smaller_state = TRUE)
}
\arguments{
\item{cn}{Copy number data}

\item{prob01}{Probability of transitioning from state CN0 to state CN1+}

\item{prob10}{Probability of transitioning from state CN1+ to state CN0}

\item{mean0}{Mean of state 0}

\item{mean1}{Mean of state 1}

\item{python_exe}{Full path to python executable}
}
\value{
list of segment starts, segment ends, and the trained model
}
\description{
Segment copy number using an HMM.
}
