% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{index_to_segments}
\alias{index_to_segments}
\title{Convert vector of indices of points within segments to a list of
vectors of start and end points of those segments
Inverse of segments_to_index}
\usage{
index_to_segments(v)
}
\description{
Convert vector of indices of points within segments to a list of
vectors of start and end points of those segments
Inverse of segments_to_index
}
