% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_distance_matrix_functions.R
\name{interpolate_peak}
\alias{interpolate_peak}
\title{Interpolates the value of the ASCAT distance matrix at (psi, rho)}
\usage{
interpolate_peak(dm, psi, rho)
}
\arguments{
\item{dm}{Matrix Distance matrix produced by ASCAT, or calculated from
an ASCAT segmentation object using the function \code{make_distance_matrix}}
}
\description{
Interpolates the value of the ASCAT distance matrix at (psi, rho)
}
