% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplot.R
\name{karyoplot_devil_genome}
\alias{karyoplot_devil_genome}
\title{Build a GRanges object describing the devil genome}
\usage{
karyoplot_devil_genome()
}
\description{
Build a GRanges object describing the devil genome
}
