% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplot.R
\name{karyoplot_dog_genome}
\alias{karyoplot_dog_genome}
\title{Build a GRanges object describing the dog genome}
\usage{
karyoplot_dog_genome()
}
\description{
Build a GRanges object describing the dog genome
}
