% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplot.R
\name{karyoplot_load_baflogr}
\alias{karyoplot_load_baflogr}
\title{Load the two tumour files that ASCAT worked on}
\usage{
karyoplot_load_baflogr(filename.baf, filename.logr)
}
\description{
Load the two tumour files that ASCAT worked on
}
