% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplot.R
\name{karyoplot_load_segments}
\alias{karyoplot_load_segments}
\title{Load the segmentation that ASCAT produced}
\usage{
karyoplot_load_segments(filename)
}
\description{
Load the segmentation that ASCAT produced
}
