% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflected_beta_distribution.R
\name{ll_reflbeta}
\alias{ll_reflbeta}
\title{Log-likelihood of a sample conditioned on these parameters, assuming reflected beta model}
\usage{
ll_reflbeta(params, data)
}
\description{
Log-likelihood of a sample conditioned on these parameters, assuming reflected beta model
}
