% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_symmetrical_beta_mixture_distribution.R
\name{lldoublesbeta}
\alias{lldoublesbeta}
\title{Log-likelihood of a sample conditioned on these parameters, assuming symmetrical beta mixture model}
\usage{
lldoublesbeta(xs, a1, b1, a2, b2, w1, w2, w3)
}
\description{
Log-likelihood of a sample conditioned on these parameters, assuming symmetrical beta mixture model
}
