% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrical_beta_mixture_distribution.R
\name{llsbeta}
\alias{llsbeta}
\title{Log-likelihood of a sample conditioned on these parameters, assuming symmetrical beta mixture model}
\usage{
llsbeta(xs, a, b, w)
}
\description{
Log-likelihood of a sample conditioned on these parameters, assuming symmetrical beta mixture model
}
