% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_odds_ratio_computation.R
\name{log_odds_ratio_variance}
\alias{log_odds_ratio_variance}
\title{Calculate the variance of the log odds ratio (logOR) of tumour and reference allele counts.
Formulation is from Facets package.}
\usage{
log_odds_ratio_variance(tumour_alt, tumour_ref, host_alt, host_ref)
}
\arguments{
\item{tumour_alt}{- count of tumour alt reads at a site}

\item{tumour_ref}{- count of tumour ref reads at a site}

\item{host_alt}{- count of host alt reads at a site}

\item{host_ref}{- count of host ref reads at a site}
}
\description{
Calculate the variance of the log odds ratio (logOR) of tumour and reference allele counts.
Formulation is from Facets package.
}
