% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood for deconvoluting host and tumour read counts}
\usage{
loglik(PHa, PHb, n, v, r, pur, a, b)
}
\arguments{
\item{PHa}{= parameter probability read is Host given read is A (P(H|A))}

\item{PHb}{= parameter probability read is Host given read is B (P(H|B))}

\item{n}{= normal read count in tumour}

\item{v}{= variant read count in tumour}

\item{r}{= ratio of tumour depth to host depth (Note: NOT logR)}

\item{pur}{= tumour purity}

\item{a}{= Expected number of host reads}

\item{b}{= Expected number of tumour reads}

\item{logr}{= log ratio of tumour depth to host depth}
}
\description{
Log-likelihood for deconvoluting host and tumour read counts
}
\examples{
loglik(0.46, 0.42, 90, 15, exp(0.03670243), 0.59, 71, 8)
# [1] 3.356328
}
