% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_correction.R
\name{logRCorrector}
\alias{logRCorrector}
\title{Creates an object to do GATK logR standardization and denoising using a panel of normals.
This is simplified compared to the GATK procedure, as it doesn't apply any filters to the
data, except for an optional winsorization. As such, it is advisable to filter out rows of
the host matrix that have very low or high coverage, or which have row-median of zero.}
\description{
Creates an object to do GATK logR standardization and denoising using a panel of normals.
This is simplified compared to the GATK procedure, as it doesn't apply any filters to the
data, except for an optional winsorization. As such, it is advisable to filter out rows of
the host matrix that have very low or high coverage, or which have row-median of zero.

Creates an object to do GATK logR standardization and denoising using a panel of normals.
This is simplified compared to the GATK procedure, as it doesn't apply any filters to the
data, except for an optional winsorization. As such, it is advisable to filter out rows of
the host matrix that have very low or high coverage, or which have row-median of zero.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{row_medians}}{Row medians of the input host read count matrix}

\item{\code{U}}{Result of Singular Value Decomposition of processed host read counts,
used for denoising tumour logR.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{logRCorrector$new()}}
\item \href{#method-standardize}{\code{logRCorrector$standardize()}}
\item \href{#method-denoise}{\code{logRCorrector$denoise()}}
\item \href{#method-clone}{\code{logRCorrector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor. Builds new object using host read count matrix as input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logRCorrector$new(host_matrix, winsorization_quantile = 0.001)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host_matrix}}{Matrix of read counts, with samples as columns.}

\item{\code{winsorization_quantile}}{Values outside the quantile range set here
(q, 1-q) are clamped to the values at those quantiles.
Can be set to 0 to disable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardize"></a>}}
\if{latex}{\out{\hypertarget{method-standardize}{}}}
\subsection{Method \code{standardize()}}{
Calculate standardized (not denoised) logR from tumour read counts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logRCorrector$standardize(tumour_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tumour_matrix}}{Matrix of tumour read counts, samples in columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-denoise"></a>}}
\if{latex}{\out{\hypertarget{method-denoise}{}}}
\subsection{Method \code{denoise()}}{
Calculate denoised logR from tumour read counts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logRCorrector$denoise(tumour_matrix, num_svs, randomized = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tumour_matrix}}{Matrix of tumour read counts, samples in columns}

\item{\code{num_svs}}{Number of singular values to use for denoising.}

\item{\code{randomized}}{Set to TRUE to use a random selection of
singular values. Set to FALSE to use the largest singular values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logRCorrector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
