% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_filters.R
\name{make_breakpoint_test_lookup_table}
\alias{make_breakpoint_test_lookup_table}
\title{Makes a lookup table to speed up the adjacent segments filter.}
\usage{
make_breakpoint_test_lookup_table(samples_list, compute_pval = FALSE)
}
\description{
Makes a lookup table to speed up the adjacent segments filter.
}
