% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_distance_matrix_functions.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{Make a distance matrix directly from the segmentation data
(this is how runASCAT does it)}
\usage{
make_distance_matrix(ascat_data)
}
\arguments{
\item{ascat_data}{ASCAT segmentation object}
}
\description{
Make a distance matrix directly from the segmentation data
(this is how runASCAT does it)
}
