% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximisation_beta_mixture_model.R
\name{max_relerror}
\alias{max_relerror}
\title{Return the maximum scaled absolute difference between vectors
va and vb}
\usage{
max_relerror(va, vb)
}
\description{
Return the maximum scaled absolute difference between vectors
va and vb
}
