% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mirror}
\alias{mirror}
\title{Reflect values in v around 0.5
(assumes 0 <= val <= 1, for all val in v)}
\usage{
mirror(v)
}
\arguments{
\item{v}{A vector of numbers in the interval [0, 1]}
}
\description{
Reflect values in v around 0.5
(assumes 0 <= val <= 1, for all val in v)
}
\examples{
mirror(c(0.1, 0.2, 0.6, 0.7, 0.4, 0.8))
# [1] 0.1 0.2 0.4 0.3 0.4 0.2
}
