% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspcf.R
\name{modded_aspcf}
\alias{modded_aspcf}
\title{If the penalty is set, then the segmentation is done to that penalty,
regardless of how many segments it ends up with}
\usage{
modded_aspcf(
  ASCATobj,
  selectsamples = 1:length(ASCATobj$samples),
  ascat.gg = NULL,
  penalty = 25,
  out.dir = ".",
  out.prefix = ""
)
}
\description{
If the penalty is set, then the segmentation is done to that penalty,
regardless of how many segments it ends up with
}
