% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{no_touching}
\alias{no_touching}
\title{Finds adjacent segments within a sample that have the same
end position and start position, and moves the end position
back by one unit.}
\usage{
no_touching(segmentation)
}
\description{
Finds adjacent segments within a sample that have the same
end position and start position, and moves the end position
back by one unit.
}
