% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{pack_ascat}
\alias{pack_ascat}
\title{An ASCAT segmentation object is a list of length 14 containing:
1  $Tumor_LogR
2  $Tumor_BAF
3  $Tumor_LogR_segmented
4  $Tumor_BAF_segmented
5  $Germline_LogR
6  $Germline_BAF
7  $SNPpos
8  $ch
9  $chr
10 $chrs
11 $samples
12 $gender
13 $sexchromosomes
14 $failedarrays}
\usage{
pack_ascat(data, segmentation)
}
\description{
An ASCAT segmentation object is a list of length 14 containing:
1  $Tumor_LogR
2  $Tumor_BAF
3  $Tumor_LogR_segmented
4  $Tumor_BAF_segmented
5  $Germline_LogR
6  $Germline_BAF
7  $SNPpos
8  $ch
9  $chr
10 $chrs
11 $samples
12 $gender
13 $sexchromosomes
14 $failedarrays
}
