% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{pack_chrs}
\alias{pack_chrs}
\title{Output: 2 * list [length=Nchrom], where list[[K]]
is the sequence 1:N, where N is the number of datapoints
coming from chromosome K
obj$ch and obj$chr are identical
obj$chrs is a character vector of each K}
\usage{
pack_chrs(data)
}
\description{
Output: 2 * list [length=Nchrom], where list[[K]]
is the sequence 1:N, where N is the number of datapoints
coming from chromosome K
obj$ch and obj$chr are identical
obj$chrs is a character vector of each K
}
