% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{pack_gender}
\alias{pack_gender}
\title{Output: character vector [length=Nsamples]
Sex of each sample ('XX' or 'XY' if sexchromosomes are 'X', 'Y')}
\usage{
pack_gender(sample_data_list)
}
\description{
Output: character vector [length=Nsamples]
Sex of each sample ('XX' or 'XY' if sexchromosomes are 'X', 'Y')
}
