% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{pack_germline_logr}
\alias{pack_germline_logr}
\title{Output: data.frame [Nsites * Nsamples]}
\usage{
pack_germline_logr(sample_data_list)
}
\description{
Output: data.frame [Nsites * Nsamples]
}
