% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{pack_snppos}
\alias{pack_snppos}
\title{Output: data.frame [NSites * 2]
colnames are "CHROM" and "POS"}
\usage{
pack_snppos(data)
}
\description{
Output: data.frame [NSites * 2]
colnames are "CHROM" and "POS"
}
