% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{pack_tumour_baf_segmented}
\alias{pack_tumour_baf_segmented}
\title{Output: list [length=Nsamples] of matrix [NHetSites * 1]
rownames of matrix must correspond to the indices of SNPpos
that point to heterozygous sites}
\usage{
pack_tumour_baf_segmented(sample_data_list, segmentation)
}
\description{
Output: list [length=Nsamples] of matrix [NHetSites * 1]
rownames of matrix must correspond to the indices of SNPpos
that point to heterozygous sites
}
