% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_ascat_object.R
\name{pack_tumour_logr_segmented}
\alias{pack_tumour_logr_segmented}
\title{Output: matrix [Nsites * Nsamples]}
\usage{
pack_tumour_logr_segmented(sample_data_list, segmentation)
}
\description{
Output: matrix [Nsites * Nsamples]
}
