% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_filters.R
\name{pairwise_filter}
\alias{pairwise_filter}
\title{Runs the pairwise filter on the samples given in comp_table. This filter compares
the copy number call for each sample to the call in the majority of samples. If
the copy number is different to the majority, then it must have a mean total copy
number that differs from the majority mean by at least `threshold`.
See documentation for `multiple_comparison_check` for details.}
\usage{
pairwise_filter(comp_table, threshold = 0.8, majority = 0.8)
}
\arguments{
\item{comp_table}{Table comparing the segment means for each sample in the set.
Can be created using `make_comparison_table(samples_list)`}

\item{threshold}{float}

\item{majority}{float}
}
\description{
Runs the pairwise filter on the samples given in comp_table. This filter compares
the copy number call for each sample to the call in the majority of samples. If
the copy number is different to the majority, then it must have a mean total copy
number that differs from the majority mean by at least `threshold`.
See documentation for `multiple_comparison_check` for details.
}
