% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{pcf}
\alias{pcf}
\title{Implementation of PCF algorithm from
Copynumber: Efficient algorithms for
single and multi-track copy number
segmentation
(not very efficient version)
example:
data = c(rnorm(495, 0, 0.1), rnorm(10, 0.5, 0.1), rnorm(495, 0, 0.1), rnorm(100, 1, 0.1), rnorm(1000, 1, 0.1))
pcf(data)}
\usage{
pcf(y, gamma = 40)
}
\description{
Implementation of PCF algorithm from
Copynumber: Efficient algorithms for
single and multi-track copy number
segmentation
(not very efficient version)
example:
data = c(rnorm(495, 0, 0.1), rnorm(10, 0.5, 0.1), rnorm(495, 0, 0.1), rnorm(100, 1, 0.1), rnorm(1000, 1, 0.1))
pcf(data)
}
