% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_symmetrical_beta_mixture_distribution.R
\name{pdoublesbeta}
\alias{pdoublesbeta}
\title{PDF function for symmetrical beta mixture}
\usage{
pdoublesbeta(q, a1, b1, a2, b2, w1, w2, w3)
}
\description{
PDF function for symmetrical beta mixture
}
