% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_distance_matrix_functions.R
\name{peakfind}
\alias{peakfind}
\title{Find peaks in the ASCAT sunrise plot data.}
\usage{
peakfind(dm, nbr_size = 5)
}
\arguments{
\item{dm}{= Distance matrix from ascat_result$distance_matrix}

\item{nbr_size}{= Neighbourhood size for peakfinding (e.g. 5x5 matrix centred on peak)}
}
\description{
Find peaks in the ASCAT sunrise plot data.
}
