% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_maximisation_beta_mixture_model.R
\name{plot_estimation_step}
\alias{plot_estimation_step}
\title{Plot a single iteration of the mixture model estimation procedure of `estimate_mixture`}
\usage{
plot_estimation_step(data, params, pi, step)
}
\arguments{
\item{"data"}{The observations being fitted}

\item{"params"}{Current Beta component `a` and `b` parameters}

\item{"pi"}{Component mixture weights}

\item{"step"}{Current iteration number}
}
\description{
Plot a single iteration of the mixture model estimation procedure of `estimate_mixture`
}
