% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{prefilter}
\alias{prefilter}
\title{Applies a prefilter to a PCF segmentation}
\usage{
prefilter(dta, seg, pthresh = 0.3, med_diff = 0.5)
}
\arguments{
\item{dta}{The data that was segmented}

\item{seg}{The result of segmenting dta}
}
\value{
A filtered segmentation
}
\description{
Applies a prefilter to a PCF segmentation
}
