% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{prob_read_came_from_tumour}
\alias{prob_read_came_from_tumour}
\title{Probability that a read came from the tumour}
\usage{
prob_read_came_from_tumour(logr, purity, ploidy, host_ploidy)
}
\arguments{
\item{logr}{Tumour sample logR}

\item{purity}{Tumour sample purity}

\item{ploidy}{Tumour ploidy (ploidy of pure tumour)}

\item{host_ploidy}{Host ploidy estimate (almost always will be 2)
Derivation:
Expected proportion of tumour reads in a mixed sample, where Nt is
underlying tumour copy number state, Nh is host copy state, and p
is purity:
  P(read=Tumour) = p * Nt / (p * Nt + (1 - p) * Nh) (1)

Tumour copynumber state is estimated as:
  Nt = (R * Nh * (p*ψt + (1-p) * ψh) - ψh * (1 - p) * Nh)
       --------------------------------------------------  (2)
                           p * ψh
Therefore,
 p * Nt = (R * Nh * (p*ψt + (1-p) * ψh) - ψh * (1 - p) * Nh)
          --------------------------------------------------  (3)
                            ψh
And,
 p * Nt + (1 - p) * Nh
        = (R * Nh * (p*ψt + (1-p) * ψh)
          -----------------------------   (4)
                      ψh
Substitute (3) and (4) into (1) to obtain the result.}
}
\description{
Probability that a read came from the tumour
}
