% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrical_beta_mixture_distribution.R
\name{qsbeta}
\alias{qsbeta}
\title{Quantile (CDF) function for symmetrical beta mixture}
\usage{
qsbeta(p, a, b, w)
}
\description{
Quantile (CDF) function for symmetrical beta mixture
}
