% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_copynumber_to_data.R
\name{quick_fit_total_copy_number}
\alias{quick_fit_total_copy_number}
\title{Find the best fitting total copy number state for a vector of logR values,
for the given purity and ploidy values}
\usage{
quick_fit_total_copy_number(logr, purity, ploidy)
}
\description{
Find the best fitting total copy number state for a vector of logR values,
for the given purity and ploidy values
}
