% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logr_computation.R
\name{quick_logr}
\alias{quick_logr}
\title{Quick calculation of logR from host and tumour coverage data
Example:
# get reference median host information:
host_index = ... (index of host data columns)
median_host <- data[, get_median_host(.SD), .SDcols = host_index]}
\usage{
quick_logr(m, rowmed)
}
\description{
# Compute logR for tumour columns
tumour_columns = ... (names of tumour columns [result written here])
tumour_index = ... (index of tumour coverage data columns)
stopifnot(length(tumour_columns) == length(tumour_index))
data[, (tumour_columns) := quick_logr(.SD, median_host),
     .SDcols = tumour_index]
}
