% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{rcpp_module_find_mle}
\alias{rcpp_module_find_mle}
\title{Helper function to find the maximum likelihood estimate of P(H|A) and P(H|B)}
\usage{
rcpp_module_find_mle(
  tumour_ref,
  tumour_alt,
  host_ref,
  host_alt,
  logr,
  purity,
  hess = FALSE
)
}
\description{
Helper function to find the maximum likelihood estimate of P(H|A) and P(H|B)
}
