% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resegment_logr}
\alias{resegment_logr}
\title{Returns segmented logR using pcf from the copynumber package
Intended for finishing off already segmented data, so ideally
the logr vector will be fairly short, and the penalty fairly
high}
\usage{
resegment_logr(chr, pos, logr, penalty = 100, kmin = 5)
}
\arguments{
\item{chr}{(character) Chromosome name}

\item{pos}{(integer vector) Genome positions}

\item{logr}{(real vector) logR values at each position}

\item{penalty}{(real) amount to penalise the PCF algorithm
creating a new segment (higher values -> fewer segments)}

\item{kmin}{(int) minimum number of data points in aa segment}
}
\description{
Returns segmented logR using pcf from the copynumber package
Intended for finishing off already segmented data, so ideally
the logr vector will be fairly short, and the penalty fairly
high
}
