% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrical_beta_mixture_distribution.R
\name{rsbeta}
\alias{rsbeta}
\title{Random variates function for symmetrical beta mixture}
\usage{
rsbeta(n, a, b, w)
}
\description{
Random variates function for symmetrical beta mixture
}
