% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{segments_to_index}
\alias{segments_to_index}
\title{Convert vectors of start and end points of segments to
vector of indices of points within those segments
Inverse of index_to_segments}
\usage{
segments_to_index(starts, ends)
}
\description{
Convert vectors of start and end points of segments to
vector of indices of points within those segments
Inverse of index_to_segments
}
