% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{simulated_logr}
\alias{simulated_logr}
\title{Simulate logr according to certain parameters}
\usage{
simulated_logr(samplesize, depth, tn, hn, purity, tploidy = 2,
  hploidy = 2)
}
\description{
Simulate logr according to certain parameters
}
