% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{simulated_vaf}
\alias{simulated_vaf}
\title{Simulate VAF according to certain parameters}
\usage{
simulated_vaf(samplesize, depth, ta, tb, ha, hb, purity, tploidy = 2,
  hploidy = 2, balance = 0)
}
\description{
Simulate VAF according to certain parameters
}
