% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_filters.R
\name{subclonal_search_pipeline}
\alias{subclonal_search_pipeline}
\title{1) Take a subset of samples
2) Scan subset for potential new subclonal breakpoints
3) Update the calls table for the subset (possibly augmented subset) using the new breakpoints
4) Median-Shift filter the calls
5) Adjacent-Segments filter the calls (optional)}
\usage{
subclonal_search_pipeline(
  samples_to_search,
  samples_to_update,
  input_calls,
  lookup_table,
  samples_list,
  segmentation_dt,
  effect_size_threshold = 0.3,
  update_method = "subclonal",
  do_adjacent_segments_filter = FALSE
)
}
\description{
1) Take a subset of samples
2) Scan subset for potential new subclonal breakpoints
3) Update the calls table for the subset (possibly augmented subset) using the new breakpoints
4) Median-Shift filter the calls
5) Adjacent-Segments filter the calls (optional)
}
