% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_filters.R
\name{tost_prefilter}
\alias{tost_prefilter}
\title{Uses TOST (two one-sided tests) to check if the copy number data within two adjacent segments
differs by at least `tol`, at a significance level of `alpha`.
If this is true, then the breakpoint will be retained in the "approved_breaks" list, otherwise
it is removed from the segmentation, and the segmentation is rebuilt.}
\usage{
tost_prefilter(dta, seg, tol = 0.3, alpha = 0.05)
}
\description{
Uses TOST (two one-sided tests) to check if the copy number data within two adjacent segments
differs by at least `tol`, at a significance level of `alpha`.
If this is true, then the breakpoint will be retained in the "approved_breaks" list, otherwise
it is removed from the segmentation, and the segmentation is rebuilt.
}
