% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_filters.R
\name{update_calls}
\alias{update_calls}
\title{Updates the copynumber calls in calls_dt for the subset of samples in samplenames,
using the provided set of breakpoints}
\usage{
update_calls(
  calls_dt,
  samples_list,
  samplenames,
  breakpoints,
  segmentation_dt,
  use_tetraploid_copynumber = FALSE,
  method = "subclonal"
)
}
\description{
Updates the copynumber calls in calls_dt for the subset of samples in samplenames,
using the provided set of breakpoints
}
