% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{v.convert_to_het_host}
\alias{v.convert_to_het_host}
\title{Uses \code{deconvolute} to calculate adjusted tumour vaf as it would be if all host contamination were heterozygous
(vectorised over indicated parameters)}
\usage{
v.convert_to_het_host(tumour_ref, tumour_alt, host_ref, host_alt, logr,
  purity)
}
\arguments{
\item{tumour_ref}{(vectorised) - number of REF reads observed in host-contaminated tumour, at current site}

\item{tumour_alt}{(vectorised) - number of ALT reads observed in host-contaminated tumour, at current site}

\item{host_ref}{(vectorised) - number of REF reads observed in matched host, at current site}

\item{host_alt}{(vectorised) - number of ALT reads observed in matched host, at current site}

\item{logr}{(vectorised) - log of ratio between tumour coverage and matched host coverage, at current site}

\item{purity}{(NOT vectorised) - Estimate of tumour purity (aka aberrant cell count)}
}
\value{
vector of VAF values
}
\description{
Uses \code{deconvolute} to calculate adjusted tumour vaf as it would be if all host contamination were heterozygous
(vectorised over indicated parameters)
}
