% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat_functions.R
\name{write_ascat_data}
\alias{write_ascat_data}
\title{Write the 4 required files for ascat}
\usage{
write_ascat_data(data, filenames)
}
\arguments{
\item{data}{- contains data needed by ascat}

\item{filenames}{- list of 4 filenames, accessed as: tumour_logr, normal_logr, tumour_baf, normal_baf}
}
\description{
Write the 4 required files for ascat
}
