% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{.estimate_contingency_table}
\alias{.estimate_contingency_table}
\title{Fast deconvolution of read count contingency table}
\usage{
.estimate_contingency_table(total_readdepth, alt_readdepth, logr, hvaf, purity)
}
\arguments{
\item{total_readdepth}{Tumour sample total read depth}

\item{alt_readdepth}{Tumour sample alt read depth}

\item{logr}{Tumour sample logR}

\item{hvaf}{Host sample VAF}

\item{purity}{Estimated purity of tumour sample
       |  Ref  |  Alt  |
-------|-------|-------|-------
 Host  |  (a)  |  (b)  |  (c)
Tumour |  (d)  |  (e)  |  (f)
-------|-------|-------|-------
       |   R   |   A   |   T
   |  Ref  |  Alt  |
---|-------|-------|-------
 H |  K-L  |   L   |   K
 T |T-K-A+L|  A-L  |  T-K
---|-------|-------|-------
   |   R   |   A   |   T}
}
\description{
Fast deconvolution of read count contingency table
}
